requirements/base.txt
Django==1.7.1
psycopg2==2.5.3
django-facebook==6.0.1
Pillow==2.9.0
Willow==0.2.2
mock==1.0.1
paypalrestsdk==1.4.1
pytz==2014.7
django-compressor==1.4
django-storages==1.1.7
django-imagekit==3.0
django-widget-tweaks==1.3
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
boto==2.34.0
celery==3.1.17
django-celery==3.1.16
beautifulsoup4==4.4.0
WebTest==2.0.18
django-webtest==1.7.7
factory_boy==2.5.2
django-ckeditor==4.4.8
wagtail==1.0b2
wagtailsettings==0.4.1
django-sekizai==0.8.1
django-sesame==1.1
simple-salesforce>=0.68
stripe==1.29.1
PyYAML==3.11
six==1.9.0
django-dotenv==1.3.0

python-social-auth==0.2
  python-openid==2.2.5
  oauthlib==1.0.3
  PyJWT==1.4.0
  requests==2.9.1
  requests-oauthlib==0.6.0
  # six (required by others)
